Add 9base recipes

+++ suckless/install.sh
@@ -2,10 +2,12 @@
 name="$(basename $PWD)"
 cp -rf src build
 pushd build
-cp -rf ../config.h .
+[[ -f ../config.h ]] && cp -rf ../config.h .
 for patch in ../*.diff; do
     patch -p1 < $patch
 done
-make install PREFIX=$PREFIXPATH/$name &&
-popd &&
-rm -rf build
+if command -v musl-gcc &>/dev/null; then
+    make install PREFIX=$PREFIXPATH/$name CC=musl-gcc && popd && rm -rf build
+else
+    make install PREFIX=$PREFIXPATH/$name && popd && rm -rf build
+fi
